everything seems to work now.
