+++ rustbot_discord/Cargo.toml
@@ -33,8 +33,3 @@ songbird = { version = "0.3.2", features = ["yt-dlp"] }
 tokio = { version = "1.21.2", features = ["macros", "rt-multi-thread"] }
 tokio-postgres = "0.7.10"
 
-
-
-[profile.dev]
-incremental = true
-codegen-units = 16
