+++ rustbot_web/Cargo.toml
@@ -34,7 +34,3 @@ tokio = { version = "1.21.2", features = ["macros", "rt-multi-thread"] }
 tokio-postgres = "0.7.10"
 
 
-
-[profile.dev]
-incremental = true
-codegen-units = 16
